Move global snackbar store
