+++ src/stores/AppStore.js
@@ -11,7 +11,7 @@ import FinalizeStore from './activitiesStores/Finalize';
 import WithdrawStore from './activitiesStores/Withdraw';
 import ActivityHistoryStore from './activitiesStores/ActivityHistoryStore';
 import WalletStore from './wallet/WalletStore';
-import GlobalSnackbarStore from './components/GlobalSnackbarStore';
+import GlobalSnackbarStore from '../components/GlobalSnackbar/store';
 import SelectAddressDialogStore from './components/SelectAddressDialogStore';
 import WalletUnlockDialogStore from './components/WalletUnlockDialogStore';
 import PendingTxsSnackbarStore from '../components/PendingTxsSnackbar/store';
